requirements.txt
Flask==1.1.2
Flask-Cors==3.0.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.5
six==1.15.0
Werkzeug==1.0.1
pillow
numpy
opencv-python
tensorflow==2.19.0
keras